style: move component props in the same file
