--- app/lib/matador/components/job-data/JobDataPanelProps.ts
@@ -1,7 +1,0 @@
-interface JobDataPanelProps {
-  parent?: any;
-  opts?: any;
-  data?: any;
-}
-
-export default JobDataPanelProps;
