--- app/lib/matador/components/job-data/JobDataProps.ts
@@ -1,6 +1,0 @@
-interface JobDataProps {
-  title: string;
-  json: any;
-}
-
-export default JobDataProps;
