--- app/lib/matador/components/job-info/JobInfoPanelProps.tsx
@@ -1,11 +1,0 @@
-interface JobInfoPanelProps {
-  id: string;
-  queueName: string;
-  jobName: string;
-  timestamp: Date;
-  processedOn?: Date;
-  finishedOn?: Date;
-  attemptsMade?: number;
-}
-
-export default JobInfoPanelProps;
