--- app/lib/matador/components/job-info/JobInfoProps.tsx
@@ -1,9 +1,0 @@
-import { MantineColor } from "@mantine/core";
-
-interface JobInfoProps {
-  infoName: string;
-  infoValue: string;
-  color?: MantineColor;
-}
-
-export default JobInfoProps;
