--- app/lib/matador/components/job-result/JobResultPanelProps.ts
@@ -1,7 +1,0 @@
-interface JobResultPanelProps {
-  returnvalue: any | undefined;
-  failedReason: string | undefined;
-  stacktrace: string[] | undefined;
-}
-
-export default JobResultPanelProps;
