--- app/lib/matador/components/job-result/JobResultProps.ts
@@ -1,7 +1,0 @@
-interface JobResultProps {
-  label: string;
-  description: string;
-  value: string;
-}
-
-export default JobResultProps;
