--- app/lib/matador/components/jobs-table/JobStatusFilterModalProps.ts
@@ -1,11 +1,0 @@
-import { JobStatus } from "../../types/JobStatus";
-
-interface JobStatusFilterModalProps {
-  opened: boolean;
-  repeatedJobs: boolean;
-  onClose: () => void;
-  selected: JobStatus[];
-  onApply: (selected: JobStatus[]) => void;
-}
-
-export default JobStatusFilterModalProps;
