--- app/lib/matador/components/jobs-table/JobsStatusHeaderProps.ts
@@ -1,6 +1,0 @@
-interface JobStatusHeaderProps {
-  isFiltered: boolean;
-  onFilterClick: () => void;
-}
-
-export default JobStatusHeaderProps;
