--- app/lib/matador/components/jobs-table/JobsTableProps.tsx
@@ -1,12 +1,0 @@
-import { BullJob } from "~/lib/matador/index.server";
-import { JobStatus } from "../../types/JobStatus";
-
-interface JobsTableProps {
-  jobs: BullJob[];
-  queueName: string;
-  repeatJobs?: boolean;
-  statusSelected: JobStatus[];
-  onStatusesSelected: (statuse: JobStatus[]) => void;
-}
-
-export default JobsTableProps;
