--- app/lib/matador/components/jobs-table/JobsTableRowProps.ts
@@ -1,9 +1,0 @@
-import { BullJob } from "~/lib/matador/index.server";
-
-interface JobsTableRowProps {
-  job: BullJob;
-  queueName: string;
-  repeatJob: boolean;
-}
-
-export default JobsTableRowProps;
