+++ app/lib/matador/components/navbar/Navlink.tsx
@@ -7,7 +7,14 @@ import {
   MantineColor,
 } from "@mantine/core";
 import { NavLink } from "react-router-dom";
-import NavlinkProps from "./NavlinkProps";
+
+export interface NavlinkProps {
+  icon: React.ReactNode;
+  color: MantineColor;
+  label: string;
+  href: string;
+  onClick: () => void;
+}
 
 const Navlink = ({ icon, color, label, href, onClick }: NavlinkProps) => {
   return (
