--- app/lib/matador/components/navbar/NavlinkProps.ts
@@ -1,11 +1,0 @@
-import { MantineColor } from "@mantine/core";
-
-interface NavlinkProps {
-  icon: React.ReactNode;
-  color: MantineColor;
-  label: string;
-  href: string;
-  onClick: () => void;
-}
-
-export default NavlinkProps;
