+++ app/lib/matador/components/navbar/index.ts
@@ -1,2 +1,1 @@
 export * from "./Navbar";
-export * from "./NavbarProps";
