--- app/lib/matador/components/stat-card/StatCardProps.ts
@@ -1,9 +1,0 @@
-import { MantineColor } from "@mantine/core";
-
-interface StatCardProps {
-  title: string;
-  value: string | number;
-  color: MantineColor;
-}
-
-export default StatCardProps;
